backend/requirements.txt
numpy==1.26.4
Pillow==11.3.0
opencv-python==4.9.0.80
scikit-learn==1.5.1
tensorflow==2.20.0  # Use the latest available 2.x version to maximize compatibility
tf-keras            # CRITICAL: Re-add this package for TensorFlow 2.16+
deepface==0.0.75
google-genai==0.2.2
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0
python-multipart==0.0.6
pytesseract==0.3.10